Move GeometrySpecification to own file
