+++ src/Structures/GeometrySpecification.cpp
@@ -1,0 +1,187 @@
+#include <cstdint>
+#include <optional>
+#include <ostream>
+#include <string>
+
+#include <nameof.hpp>
+
+#include "../Enums/LineStyle.hpp"
+#include "../Enums/LineWidth.hpp"
+#include "../General.hpp"
+#include "../Parser.hpp"
+#include "GeometrySpecification.hpp"
+
+
+static FileFormatVersion predictVersion(DataStream& aDs, Parser& aParser)
+{
+    FileFormatVersion prediction = FileFormatVersion::Unknown;
+
+
+    const std::vector<FileFormatVersion> versions{
+        FileFormatVersion::A,
+        FileFormatVersion::B,
+        FileFormatVersion::C
+    };
+
+    const size_t initial_offset = aDs.getCurrentOffset();
+
+    for(const auto& version : versions)
+    {
+        bool found = true;
+
+        GeometrySpecification geo_spec;
+
+        try
+        {
+            aParser.parseGeometrySpecification(version);
+        }
+        catch(...)
+        {
+            found = false;
+        }
+
+        aDs.setCurrentOffset(initial_offset);
+
+        if(found)
+        {
+            prediction = version;
+            break;
+        }
+    }
+
+    if(prediction == FileFormatVersion::Unknown)
+    {
+        // Set to previous default value
+        // s.t. tests not fail
+        prediction = FileFormatVersion::C;
+    }
+
+    return prediction;
+}
+
+
+GeometrySpecification Parser::parseGeometrySpecification(FileFormatVersion aVersion)
+{
+    spdlog::debug(getOpeningMsg(__func__, mDs.getCurrentOffset()));
+
+    const std::optional<FutureData> thisFuture = getFutureData();
+
+    // Predict version
+    if(aVersion == FileFormatVersion::Unknown)
+    {
+        aVersion = predictVersion(mDs, *this);
+        // spdlog::info("Predicted version {} in {}", aVersion, __func__);
+    }
+
+    GeometrySpecification obj;
+
+    obj.name = mDs.readStringLenZeroTerm();
+
+    mDs.assumeData({0x00, 0x00, 0x00}, std::string(__func__) + " - 0"); // Unknown but probably a string
+    mDs.assumeData({0x30}, std::string(__func__) + " - 1");
+    mDs.assumeData({0x00, 0x00, 0x00}, std::string(__func__) + " - 2"); // Unknown but probably a string
+
+    const uint16_t geometryCount = mDs.readUint16();
+    spdlog::debug("geometryCount = {}", geometryCount);
+
+    for(size_t i = 0u; i < geometryCount; ++i)
+    {
+        spdlog::debug("i of geometryCount = {}", i);
+
+        if(i > 0u)
+        {
+            if(mFileFormatVersion == FileFormatVersion::B)
+            {
+                // Structure structure = read_prefixes(3);
+                Structure structure = auto_read_prefixes();
+            }
+
+            if(mFileFormatVersion >= FileFormatVersion::B)
+            {
+                readPreamble();
+            }
+        }
+
+        GeometryStructure geometryStructure1 = ToGeometryStructure(mDs.readUint8());
+        GeometryStructure geometryStructure2 = ToGeometryStructure(mDs.readUint8());
+
+        if(geometryStructure1 != geometryStructure2)
+        {
+            throw std::runtime_error("Geometry structures should be equal!");
+        }
+
+        auto geoStruct = geometryStructure1;
+
+        readGeometryStructure(geoStruct, &obj);
+
+        // uint16_t foo = mDs.readUint8();
+        // foo = (foo << 8) | foo;
+        // geoStruct = ToGeometryStructure(foo);
+
+        // mDs.printUnknownData(40, std::string(__func__) + " - 1");
+        // readPreamble();
+
+        if(mFileFormatVersion == FileFormatVersion::A)
+        {
+            mDs.printUnknownData(8, std::string(__func__) + " - 3.5");
+        }
+    }
+
+    if(geometryCount == 0u)
+    {
+        // throw std::runtime_error("CatchMeIfYouCan");
+        // mDs.printUnknownData(6, std::string(__func__) + " - 4");
+    }
+
+    // if(geometryCount == 0u)
+    // {
+    //     mDs.printUnknownData(10, std::string(__func__) + " - 3");
+
+    //     {
+    //         GeometryStructure geoStruct;
+
+    //         Structure structure = read_type_prefix();
+    //         readConditionalPreamble(structure);
+    //         parseStructure(structure);
+
+    //         const uint16_t followingLen1 = mDs.readUint16();
+
+    //         for(size_t i = 0u; i < followingLen1; ++i)
+    //         {
+    //             std::clog << "0x" << ToHex(mDs.getCurrentOffset(), 8) << ": followingLen1 Iteration "
+    //                     << std::to_string(i + 1) << "/" << std::to_string(followingLen1) << std::endl;
+
+    //             structure = read_type_prefix();
+    //             readConditionalPreamble(structure);
+    //             // readPreamble();
+    //             parseStructure(structure);
+    //         }
+
+    //         mDs.printUnknownData(24, "foo");
+
+    //         structure = read_type_prefix();
+    //         readConditionalPreamble(structure);
+    //         parseStructure(structure);
+
+    //         structure = read_type_prefix();
+    //         readConditionalPreamble(structure);
+    //         parseStructure(structure);
+
+    //         // geoStruct = ToGeometryStructure(mDs.readUint16());
+
+    //         // uint16_t foo = mDs.readUint8();
+    //         // foo = (foo << 8) | foo;
+    //         // geoStruct = ToGeometryStructure(foo);
+    //         // readGeometryStructure(geoStruct, &obj);
+    //     }
+    // }
+
+    sanitizeThisFutureSize(thisFuture);
+
+    checkTrailingFuture();
+
+    spdlog::debug(getClosingMsg(__func__, mDs.getCurrentOffset()));
+    spdlog::info(to_string(obj));
+
+    return obj;
+}
